feat: merchant usecase 도입 2  - SearchMerchantMemberUseCase  - VatUpdateUseCase  - TipUpdateUseCase

+++ qrpay_backend/src/main/java/com/bccard/qrpay/controller/api/dtos/TipChangeReqDto.java
@@ -7,5 +7,5 @@ import lombok.ToString;
 @ToString
 public class TipChangeReqDto {
     private boolean enableTip;
-    private int tipRate;
+    private Long tipRate;
 }
