+++ qrpay_backend/src/main/java/com/bccard/qrpay/controller/api/dtos/TipTaxInfoDto.java
@@ -1,12 +1,11 @@
 package com.bccard.qrpay.controller.api.dtos;
 
-import java.math.BigDecimal;
 import lombok.Builder;
 import lombok.Getter;
 
 @Getter
 @Builder
 public class TipTaxInfoDto {
-    private BigDecimal vatRate;
-    private BigDecimal tipRate;
+    private Long vatRate;
+    private Long tipRate;
 }
